+++ app/models/start_option.rb
@@ -6,4 +6,33 @@ class StartOption < ActiveRecord::Base
     return portal_type if portal_region.nil?
     "#{portal_type} #{portal_region}"
   end
+
+  def create_options
+    command = "cd ~/RubymineProjects/SharedFunctional && git reset --hard && git pull --all --prune && git checkout #{shared_branch} && bundle install && " \
+        "cd ~/RubymineProjects/OnlineDocuments && git reset --hard && git pull --all --prune && git checkout #{docs_branch} && bundle install && " \
+        "cd ~/RubymineProjects/TeamLab && git reset --hard && git pull --all --prune && git checkout #{teamlab_branch} && bundle install && " \
+        "cd ~/RubymineProjects/TeamLabAPI2 && git reset --hard && git pull --all --prune && git checkout #{teamlab_api_branch}" \
+        "#{generate_region_command}"
+    command
+  end
+
+  def generate_region_command
+    return '' if portal_type == 'default'
+    portal_data_docs = '~/RubymineProjects/OnlineDocuments/data/portal_data.rb'
+    portal_data_teamlab = '~/RubymineProjects/TeamLab/Framework/StaticDataTeamLab.rb'
+    region_command = ' && '
+    region_command +=
+      "sed -i \\\"s/@create_portal_domain = '.*'/@create_portal_domain = '.#{portal_type}'/g\\\" #{portal_data_docs} && " \
+      "sed -i \\\"s/@create_portal_region = '.*'/@create_portal_region = '#{portal_region}'/g\\\" #{portal_data_docs} && " \
+      "sed -i \\\"s/@@portal_type = '.*'/@@portal_type = '.#{portal_type}'/g\\\" #{portal_data_teamlab} && " \
+      "sed -i \\\"s/@@server_region = '.*'/@@server_region = '#{portal_region}'/g\\\" #{portal_data_teamlab}"
+    region_command
+  end
+
+  # Check if current server options tells, that
+  # this test should be run on custom portal
+  # @return [True, False] is test should run on custom portal
+  def on_custom_portal?
+    !(portal_type == 'info' || portal_type == 'com' || portal_type == 'default')
+  end
 end