+++ Changelog.md
@@ -29,6 +29,7 @@
 * Remove unused `start-icon`
 * Reorganize a whole lot managing branches
 * Remove forced rvm sourced command from starting test
+* Remove duplicate class `ServerOptions`
 
 ### Fix
 * Use `seeds.rb` to set default `SpecLanguage` to `en-US`